Time: 1 ms (96%), Space: 44.3 MB (78.46%) - LeetHub

+++ 3251-maximum-area-of-longest-diagonal-rectangle/3251-maximum-area-of-longest-diagonal-rectangle.java
@@ -1,13 +1,13 @@
 class Solution {
     public int areaOfMaxDiagonal(int[][] dimensions) {
-            long maxArea = 0;
-            long maxDiagonal = 0;
+            int maxArea = 0;
+            int maxDiagonal = 0;
             int n = dimensions.length;
             for(int i = 0;i < n;i++) {
                  int length = dimensions[i][0];
                  int breadth = dimensions[i][1];
-                 long diagonal = (int)(Math.pow(length, 2) + Math.pow(breadth, 2));
-                 long area = length*breadth;
+                 int diagonal = (int)(Math.pow(length, 2) + Math.pow(breadth, 2));
+                 int area = length*breadth;
                  if(diagonal>maxDiagonal) {
                         maxArea = area;
                         maxDiagonal = diagonal;
@@ -16,7 +16,7 @@ class Solution {
                  }
             }
 
-            return (int)maxArea;
+            return maxArea;
             
     }
 }
